Refactored the shop code.
